+++ src/test/java/com/antique/service/ProductServiceTest.java
@@ -30,11 +30,9 @@ class ProductServiceTest {
     @Mock
     private ProductRepository productRepository; // Mock 객체
 
-    */
-/*
     * Product 클래스의 생성자가 protected로 설정되어 있어, 해당 클래스의 인스턴스를 테스트 클래스에서 직접 생성할 수 없음.
     * 따라서, Product 객체를 생성하기 위해 Product 클래스의 서브 클래스를 만들었음.
-    * *//*
+    *
 
     private static class TestProduct extends Product {
         public TestProduct(Long productId, String name, String description, int price, String status, List<String> images, String sellerNickname) {
@@ -55,11 +53,9 @@ class ProductServiceTest {
     }
 
 
-    */
-/*
     * [ 상품 전체 목록 조회 ]
     * - 상품이 존재하는 경우
-    * *//*
+    *
 
     @Test
     public void testGetAllProducts() {
@@ -79,11 +75,9 @@ class ProductServiceTest {
         assertEquals("Product 2", result.get(1).getName());
     }
 
-    */
-/*
     * [ 상품 전체 목록 조회 ]
     * - 상품이 존재하지 않는 경우
-    * *//*
+    *
 
     @Test
     public void testGetAllProducts_WhenNoProducts() {
@@ -98,10 +92,8 @@ class ProductServiceTest {
         assertEquals(0, result.size());
     }
 
-    */
-/*
      * [ 상품 카테고리별 목록 조회 ]
-     * *//*
+     *
 
     @Test
     public void testGetProductsByCategory() {
@@ -125,12 +117,10 @@ class ProductServiceTest {
         assertEquals("nickname2", result.get(1).getSellerNickName());
     }
 
-    */
-/*
      * [ 상품 상세 정보 조회 ]
      * - 상품이 존재하는 경우
      * - 상품이 존재하지 않는 경우
-     * *//*
+     *
 
     @Test
     public void testGetProductInfo() {
@@ -176,4 +166,5 @@ class ProductServiceTest {
         assertThrows(ProductNotFoundException.class, () -> productService.getProductInfo(productId));
     }
 
-}*/
+}
+*/
