+++ src/test/java/com/antique/service/UserServiceTest.java
@@ -5,6 +5,7 @@ import com.antique.domain.User;
 import com.antique.dto.user.UserRequestDTO;
 import com.antique.exception.user.UserNotFoundException;
 import com.antique.repository.UserRepository;
+import com.antique.service.user.UserService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
